requirements.txt
#
# Runtime requirements

Flask==3.1.0                    # BSD 3-Clause
Flask-SQLAlchemy==3.1.1         # BSD 3-Clause
flask-smorest==0.46.1           # MIT
Flask-Marshmallow==1.3.0        # MIT
marshmallow-sqlalchemy==1.4.2   # MIT
#Flask-OIDC~=1.4.0               # MIT

gunicorn==23.0.0                # MIT
psycopg2-binary==2.9.10         # LGPL with exceptions
